separate interface file from logic files
