+++ CMakeLists.txt
@@ -7,7 +7,9 @@ set(CMAKE_CXX_STANDARD 20)
 add_library(sqlite3 C:/sqlite3/sqlite3.c)
 
 add_executable(main 
-    main.cpp 
+    main.cpp
+    interface/customerCLI.cpp
+    interface/tellerCLI.cpp
     
     src/person.cpp 
     src/customer.cpp
@@ -15,5 +17,5 @@ add_executable(main
     src/sqlite3db.cpp
     src/utils.cpp)
 
-target_include_directories(main PUBLIC inc C:/sqlite3 inc/interface)
+target_include_directories(main PUBLIC inc C:/sqlite3 interface/inc)
 target_link_libraries(main PRIVATE sqlite3)
